+++ .gitignore
@@ -16,8 +16,8 @@ gradle-app.setting
 .gradletasknamecache
 
 # IDE
-/plume.iml
-/.idea/
+**.iml
+**/.idea/
 .project
 .settings
 .classpath
@@ -29,6 +29,6 @@ gradle-app.setting
 /sootOutput/
 /plume/sootOutput/
 /testing/build/
-/.bsp/
+**/.bsp/
 target
 /project/target/
