+++ commons/src/main/scala/com/github/plume/oss/domain/HttpResponse.scala
