+++ commons/src/main/scala/com/github/plume/oss/util/HashUtil.scala
