+++ drivers/base/build.sbt
@@ -1,0 +1,10 @@
+name := "base"
+
+libraryDependencies ++= Seq(
+  "io.shiftleft" %% "overflowdb-traversal" % Versions.overflowDb,
+  "io.shiftleft" %% "codepropertygraph"    % Versions.codePropertyGraph,
+  "org.slf4j" % "slf4j-api" % Versions.slf4j,
+  "org.scala-lang" % "scala-reflect" % scalaVersion.value,
+  "org.apache.logging.log4j" % "log4j-core" % Versions.log4j % Test,
+  "org.apache.logging.log4j" % "log4j-slf4j-impl" % Versions.log4j % Test,
+)
