+++ drivers/neo4j/build.sbt
@@ -1,0 +1,5 @@
+name := "neo4j"
+
+libraryDependencies ++= Seq(
+  "org.neo4j.driver" % "neo4j-java-driver" % Versions.neo4j
+)
