+++ drivers/tigergraph/src/test/resources/docker/TigerGraph.yml
