+++ drivers/tinkergraph/build.sbt
@@ -1,0 +1,6 @@
+name := "tinkergraph"
+
+libraryDependencies ++= Seq(
+  "org.apache.tinkerpop" % "gremlin-driver" % Versions.tinkerGraph,
+  "org.apache.tinkerpop" % "tinkergraph-gremlin" % Versions.tinkerGraph,
+)
