+++ project/Versions.scala
@@ -1,0 +1,27 @@
+object Versions {
+
+  // SAST
+  val codePropertyGraph = "1.3.542"
+  val joern             = "1.1.851"
+  val overflowDb        = "1.136"
+  val soot              = "4.3.0"
+
+  // Drivers
+  val tinkerGraph = "3.4.11"
+  val neo4j       = "4.4.5"
+
+  // Utilities
+  val apacheCodec = "1.15"
+  val apacheIo    = "2.11.0"
+  val apacheLang  = "3.12.0"
+  val sttp        = "3.5.1"
+  val jackson     = "2.13.2"
+  val scalajHttp  = "2.4.2"
+  val lz4         = "1.8.0"
+  val slf4j       = "1.7.36"
+  val log4j       = "2.17.2"
+  val logback     = "1.2.11"
+  val scalatest   = "3.2.15"
+  val circe       = "0.14.1"
+
+}
